speed problems...
please see https://groups.google.com/forum/#!topic/gitolite/mxayAy9AtSI
for a description.

TLDR: it may be a while before anything happens on this branch; most people
may find it easier to adopt the solution outlined in that URL (i.e., override
REF_OR_FILENAME_PATT and so on in the rc file, with custom patterns that
include the locale-specific alphabetics).